Abstraction rules complete

+++ src/TaxApp/SBI.java
@@ -4,7 +4,7 @@ package TaxApp;
  * Created by arash on 1/24/17.
  */
 public class SBI {
-    public int code;
+    public double code;
     public String section;
     public String department;
